requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-cors-headers==3.10.1
django-extensions==3.1.5
djangorestframework==3.13.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
